Dockerfile: add a lava user with super-user privilege
Instead of setting a hardcoded password to root,
add a lava user with super-user privilege (sudo).

Remote access to the container is still permitted
but through lava user, using SSH key(s).

Propagation of the authorized key(s) is done either by:
a. use the COPY directive and include the keys in the container
b. use -v command line argument when running the container
   (to mount a host file as a data volume)

Signed-off-by: Fathi Boudra <fathi.boudra@linaro.org>